+++ app/src/main/res/layout/item_cart.xml
@@ -5,6 +5,13 @@
     android:layout_height="150dp"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <!--
+ LIVARTC
+ Created by 변형준
+ Date: 1/19/24
+ 작업내용: 리사이클러뷰 아이템 구현
+  -->
+
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
