+++ app/src/main/res/layout/item_purchase.xml
@@ -5,6 +5,12 @@
     android:layout_width="match_parent"
     android:layout_height="150dp"
     xmlns:app="http://schemas.android.com/apk/res-auto">
+    <!--
+     LIVARTC
+     Created by 변형준
+     Date: 1/19/24
+     작업내용: 리사이클러뷰 아이템 구현
+      -->
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
@@ -34,10 +40,10 @@
                 android:text="Lorem Ipsum"
                 android:textSize="@dimen/font_medium"
                 android:layout_marginStart="20dp"
-                android:layout_marginTop="10dp"
+                android:layout_marginTop="1dp"
                 android:textStyle="bold"
-                app:layout_constraintStart_toEndOf="@+id/product_image"
-                app:layout_constraintTop_toTopOf="parent" />
+                app:layout_constraintStart_toEndOf="@id/product_image"
+                app:layout_constraintTop_toBottomOf="@id/created_at" />
 
             <TextView
                 android:id="@+id/created_at"
@@ -46,9 +52,9 @@
                 android:text="2024.01.01"
                 android:textSize="@dimen/font_small"
                 android:textStyle="bold"
-                app:layout_constraintBottom_toBottomOf="@+id/product_name"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="@+id/product_name" />
+                android:layout_marginBottom="1dp"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintEnd_toEndOf="parent" />
 
             <TextView
                 android:id="@+id/product_price"
